+++ crates/objc2/CHANGELOG.md
@@ -43,6 +43,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   ];
   ```
 
+### Fixed
+* Fixed the name of the protocol that `NSObjectProtocol` references.
+
 
 ## 0.4.1 - 2023-07-31